Explicitly select the sbcl version in .travis.yml.
In Tarvis CI, run-prove failed only in SBCL. I don't know the correct
reason. Then, although I tried to rebuild the test that is passed in
previous, it failed. The difference I found is the version of the
SBCL (but in my environment, run-prove succeed in both version). So I
try to use old version.

https://travis-ci.org/eshamster/ps-experiment/jobs/189826061