Fix package sorting in with-use-ps-pack
Previously the sort order is not enoughly considered because in many
case order of codes doesn't effect to the result in
JavaScript. However, for example, including of defstruct requires that
its parent is defined before it. So reconsidered such order in this
commit.

Note :The implementation was experimented in the following gist code.

https://gist.github.com/eshamster/4aad913f338c2852e795987798489ff6

By this commit, circular dependency of packages is prohibited. If the
constraint causes problems, it will be solved by "group-node"
implemented in the above gist code.